bai 8 : detached db and route into file private so easy to manage
