+++ db.js
@@ -1,0 +1,11 @@
+//dung tam lowdb
+const low = require('lowdb')
+const FileSync = require('lowdb/adapters/FileSync')
+const adapter = new FileSync('db.json')
+const db = low(adapter)
+
+// Set some defaults (required if your JSON file is empty)
+db.defaults({ users: [] })
+    .write()
+
+module.exports = db;
